backend_project/requirements.txt
asgiref==3.6.0
autopep8==2.0.2
black==23.3.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
coreapi==2.3.3
coreschema==0.0.4
cryptography==40.0.1
defusedxml==0.7.1
Django==4.1.7
django-colorfield==0.8.0
django-cors-headers==3.14.0
django-filter==22.1
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==4.7.2
djoser==2.1.0
flake8==6.0.0
gunicorn==20.1.0
idna==3.4
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.2
mccabe==0.7.0
mypy-extensions==1.0.0
oauthlib==3.2.2
packaging==23.1
pathspec==0.11.1
Pillow==9.4.0
platformdirs==3.2.0
psycopg2-binary==2.8.6
pycodestyle==2.10.0
pycparser==2.21
pyflakes==3.0.1
PyJWT==2.6.0
python-dotenv==1.0.0
python3-openid==3.2.0
pytz==2022.7.1
reportlab==3.6.12
requests==2.28.2
requests-oauthlib==1.3.1
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.4.0
sorl-thumbnail==12.9.0
sqlparse==0.4.3
tomli==2.0.1
typing_extensions==4.5.0
tzdata==2022.7
uritemplate==4.1.1
urllib3==1.26.15
